Create caffeine.pyw
Caffeine is started from one.vbs which is used to run the first.bat script hidden with no pop-ups or command line.
WshShell is a generic name for a powerful object that enables you to query and interact with various aspects of the Windows shell. It can display information to the user, run applications, create shortcuts, work with the Registry, and control Windows' environment variables.[11]
Here one.vbs refers to the shell object and executes first.bat while telling the shell to display nothing while executing the batch file.(See Appendix A for detailed information)
Getting in the system, first.bat is a set of simple commands that copy caffeine.exe, caffeineMeter.exe and starter.vbs into the windows directory:
“C:\Users\USER-HERE\AppData\Roaming” this is a deep directory in windows that contains application data and settings that are specific to the apps in your windows pc. This directory is not browsed by normal users 90% of the time and is safe to say hiding the malware here is a good measure.
It was tricky accomplishing this copy command because the directory has a different name in every machine because of the computer name \User-Here\ can be possibly everything. I solved this problem by researching, I found that windows uses variable names for specific directories such as %WORD%, %temp% and one of them was also %appdata%, these variable names point to the folder’s directory. So calling %appdata% in different machines points to the application data directory of that machine.
After copying the files into the directory an internet shortcut is created in the startup folder of windows. This folder is Windows default and everything inside it runs every time the computer starts. This startup folder lives inside some folders where %appdata% points to, this is the startup folder directory:
“C:\Users\USER-HERE\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\Startup”.
As we can see our variable name points up to the Roaming folder where deep inside it resides the startup folder so I added %Appdata% and as usual the other part “\Microsoft\Windows\Start Menu\Programs\Startup\”. In addition to the directory, the shortcut also needs a name which I simply called it autostart. The internet shortcut needs four parameters to be created the directory
20
where it will reside, the name, the application where does it point to and the icon.