caffeineMeter.pyw
This other part of the malware is a bit more complicated because it involves multiple libraries such as smtplib, socket, encoders, os and parts of the MIME library (MIME multipart, MIME base and MIME text). The first function is called milk I choose this name because it reassembles the milking process of a cow when extracting a file from the directory. I added this data exfiltration feature lately, before this class was only used to send the coffee.txt logs that caffeine produces.
The method takes the parameter folder which I have set to C:\\Usersbecause in this directory it can find every file that all the users of that machine have. The methods scans every folder starting from my documents, downloads, desktop, my photos, and every subfolder that exists for all of the users. This scan is what takes a little bit of time but comparing to the damage it does is worth it. The method uses the library OS which provides portable ways of using system dependent functionalities. [14]
First I created an empty string named file_list, and a loop that opens the directory we have set before as variable name folder and for every path, directory and file that exists in that directory it uses string matching to find files ending with the extension .txt, .pdf, .doc and .docx.These files types are usually where sensitive information I am looking for is found: text files pdf files or word documents. After finding these types of files it chooses only the files that start with the name passw(haven’t used password because some applications use shorted names like psw or passw), bank, statement or email. After the finding a file that matches all the conditions it gets the directory and the filename and saves them as a string to the variable I created before file_list.
This method returns file_listso when calling it we get that variable for example “C://Users/Admin/Desktop/bank_statement_uba2020.pdf”
This method can be improved further by adding the option to search for more file types/extensions and adding a faster searching algorithm.
The second method called getPcNameuses the library socket which provides access to the BSD socket interface. I have used the libraries method getHostname() and getHostByName()which in combination together return the IPV4 address and the hostname of the machine where it is currently running as a string. The method returns the string useridwhich concatenates the IP and
23
the name in one. This is later used as a subject for the email address to differentiate where the logs are coming from for example Loggs from Ernando-PC @ 112.158.126.31. [15]
I have created some variables that contain the information needed for the third and last method of the malware which are email_username, email_password, email_receiver, subject, body, filename and filename2. The sendMail method takes these variables as a parameter it uses the MIME library from the emailmodule. Firstly it creates a multipart object named msg which takes the parameters: from, to, subject and attachment which I have set equal to the variables above. To the msg object attached is MIMEtext object which defines the content and the content type in this case is the variable body and the type plainas in plain text, the method open() that is used to read a file in text format by default setting the mode parameter to rb meaning opening a file in binary format for reading.[16] [17] [18]
Second is created the MIMEbase object reading the attachment and encoding it into base64 in order to be sent, the object I called part part is then attached to the msg object, then a variable called text converts the whole msg with the partattached to string.
The connection to the SMTP server is done by firstly naming a variable server that calls the smtp library setting the servers SMTPaddress and port then startinga TSL connection to the server. After logging in with the SMTP server using the login credentials of my email account which in this case I set to the variables email_username and email_password, I use the method sendmail of the smtp library to send the mail with the other remaining parameters. These being email_username, email_receiver and the text variable which holds the whole message as a string created before, after this I call the default quit() method to close the connection.[17][18]
In the end of the file I call the method sendMail() in order to execute it with the parameters set above twice; once with the variable filename and second with the variable filename2 where the first has attached coffee.txt from caffeine.exe and the second sends the same except attached is the file that is returned from the method milk(), the sensitive data.
In the end we get two emails in the inbox giving us the victim’s key logs of the whole day everyday and their sensitive data updated because the scan runs every time the computer starts.